+++ I-Format.hpp
@@ -7,7 +7,7 @@
 using namespace std;
 
 bool IFormatChecker(string inst, bool& offset);
-void IFormat(string inst, string inst_rest, vector<pair<string, long long int> > &reg, long long int &pc, bool &pc_changed , map< long long int , long long int >  &memory, bool &err , bool&offset);
+void IFormat(string inst, string inst_rest, vector<pair<string, int> > &reg, int &pc, bool &pc_changed , map< int , int >  &memory, bool &err , bool&offset);
 void extractStrings(string input, string& beforeParentheses, string& withinParentheses);
 
 #endif
